requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
openai==1.54.3
PyMuPDF==1.23.8
pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
flower==2.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
# JWT and Authentication dependencies
PyJWT[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0